change the type of id to Long

+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvItem.java
@@ -8,6 +8,6 @@ public record CsvItem(
 		@CsvBindByName(column = "category_id") Long categoryId,
 		@CsvBindByName(column = "cost") int cost,
 		@CsvBindByName(column = "fling_power") Integer flingPower,
-		@CsvBindByName(column = "fling_effect_id") Integer flingEffectId,
+		@CsvBindByName(column = "fling_effect_id") Long flingEffectId,
 		@CsvBindByName(column = "name") String name
 ) {}
